Fix the error occurring when accessing the ctx region
BPF program arguments are placed in the ctx region.
They are aligned to 64-bits. Therefore, if an argument is
of type int, it is treated as 64-bit.

+++ my_ops.c
@@ -56,7 +56,7 @@ static bool my_ops_is_valid_access(int off, int size,
 				   const struct bpf_prog *prog,
 				   struct bpf_insn_access_aux *info)
 {
-	if (off == 0 && size == sizeof(int))
+	if (off == 0 && size == sizeof(u64))
 		return true;
 	return false;
 }
